Day 20
Part 2 required manual analysis of the input. As such, the part 2 code does *not* output an answer, but it does output information necessary to find the answer.

The trick to this one was manually analyzing the graph in the input data, looking at what feeds into `rx`, and finding patterns with that. It turns out the pattern is fairly straightforward to compute with some knowledge of the graph and the data output by this program.