Day 24
Part 2 required a hint. I was able to get much of the way there on my own, but it still required looping over an obscenely large time range (O(n^2) since I needed coordinates on two lines to base it on) or figuring out how to solve a system of equations. I ended up learning about Z3, and it turned out to be faster to solve as the system of equations using that.